Bugs fixed - Issues still
Setup error for required actions

+++ Kobt/Kobt/Timer.vb
@@ -2,13 +2,14 @@ Imports System.IO
 
 Public Class Timer
 
-    Public flag As Boolean = False
+
     Private Sub Button3_Click(sender As System.Object, e As System.EventArgs)
         Me.Close()
     End Sub
 
     Private Sub Timer_Load(sender As System.Object, e As System.EventArgs) Handles MyBase.Load
         MdiParent = Kobt
+        OvFlag = Kobt.OvFlag2
         Timer1.Start()
     End Sub
 
@@ -17,60 +18,87 @@ Public Class Timer
 
 
     End Sub
-
+    Dim SaveFlag As Boolean
+    Public OvFlag As Boolean
     Private Sub Button1_Click(sender As System.Object, e As System.EventArgs) Handles Button1.Click
+
         Dim i As Integer = NumericUpDown1.Value
         Dim k As Integer = NumericUpDown2.Value
         Label2.Text = DateAdd(DateInterval.Hour, i, TimeOfDay)
         Dim temp As Date = Label2.Text
         Dim temp2 As Date
         temp2 = DateAdd(DateInterval.Minute, k, temp)
-        Label2.Text = temp2
+        Label2.Text = Format(temp2, "HH:mm:ss")
 
-        '*****SAVE CODE TO SET BUTTON'S FUNCTION TO SAVE & SET *****
+        '*****----SAVE CODE TO SET BUTTON'S FUNCTION TO SAVE & SET---- *****
 
+        '#####Variable and path assign 
         Dim currentuser As String = Environment.UserName
         Dim strFolder = "C:\users\" + currentuser + "\documents\kobt\"
         Dim filepath As String = Path.Combine(strFolder, TextBox1.Text)
-        If String.IsNullOrEmpty(TextBox1.Text) Then
-            MessageBox.Show("Boss Name can't be empty!", "Empty Boss Name", MessageBoxButtons.OK)
-        ElseIf String.IsNullOrWhiteSpace(TextBox1.Text) Then
-            MessageBox.Show("Boss Name can't be Space!", "Space as Boss Name", MessageBoxButtons.OK)
-        Else
-            If Not System.IO.Directory.Exists(strFolder) Then
-                System.IO.Directory.CreateDirectory(strFolder)
+        '########Variable and path assign ends
+
+        '##############Cheking Directory and Save parameters Condition
+        If String.IsNullOrEmpty(TextBox1.Text) Or String.IsNullOrWhiteSpace(TextBox1.Text) Or String.IsNullOrEmpty(ComboBox1.SelectedItem) Then
+            MessageBox.Show("Boss and Zone Name can't be empty or space", "Wrong Parameters", MessageBoxButtons.OK)
+
+        ElseIf OvFlag = False And Not System.IO.Directory.Exists(strFolder) Then
+            System.IO.Directory.CreateDirectory(strFolder)
+            '##############Check ends here
+        ElseIf OvFlag = True Then
+            Dim ind As Integer = ComboBox1.SelectedIndex
+            Using sw As StreamWriter = New StreamWriter(filepath)
+                sw.WriteLine(Label2.Text)
+                sw.WriteLine(Label1.Text)
+                sw.WriteLine(ind)
+                SaveFlag = True
+                Kobt.ToolStripStatusLabel16.Text = "Saved as " + TextBox1.Text
+                Kobt.Timer2.Start()
+            End Using
+            '############ Checking File condition 
+        ElseIf OvFlag = False And System.IO.File.Exists(filepath) Then
+            Kobt.ToolStripStatusLabel16.Text = "Change Boss Name"
+            Kobt.Timer2.Start()
+            If MessageBox.Show("File " + TextBox1.Text + "  exists. Are you sure to Overwrite it?", "Warning", MessageBoxButtons.YesNo) = Windows.Forms.DialogResult.No Then
+                SaveFlag = False
+            Else
                 Dim ind As Integer = ComboBox1.SelectedIndex
-                
                 Using sw As StreamWriter = New StreamWriter(filepath)
                     sw.WriteLine(Label2.Text)
                     sw.WriteLine(Label1.Text)
                     sw.WriteLine(ind)
+                    SaveFlag = True
+                    Kobt.ToolStripStatusLabel16.Text = "Saved as " + TextBox1.Text
+                    Kobt.Timer2.Start()
                 End Using
-                'IO.File.WriteAllText(strFolder + TextBox1.Text, Label2.Text)
-                'MessageBox.Show("Saved as " + TextBox1.Text, "Succes", MessageBoxButtons.OK)
-            Else
-                If System.IO.File.Exists(filepath) Then
-                    MessageBox.Show("File " + TextBox1.Text + "  exists. (Input the zone after the Boss name)", "Warning")
-
-                Else
-                    Dim ind As Integer = ComboBox1.SelectedIndex
-                    Using sw As StreamWriter = New StreamWriter(filepath)
-                        sw.WriteLine(Label2.Text)
-                        sw.WriteLine(Label1.Text)
-                        sw.WriteLine(ind)
-                    End Using
-                    'IO.File.WriteAllText(strFolder + TextBox1.Text, Label2.Text)
-                    'MessageBox.Show("Saved as " + TextBox1.Text, "Succes", MessageBoxButtons.OK) 
-                End If
-
             End If
+
+            '####### File Check Ends here
+
+            '####### Actions If file not exists
+
+        ElseIf OvFlag = False And Not System.IO.File.Exists(filepath) Then
+            Dim ind As Integer = ComboBox1.SelectedIndex
+            Using sw As StreamWriter = New StreamWriter(filepath)
+                sw.WriteLine(Label2.Text)
+                sw.WriteLine(Label1.Text)
+                sw.WriteLine(ind)
+                SaveFlag = True
+                Kobt.ToolStripStatusLabel16.Text = "Saved as " + TextBox1.Text
+                Kobt.Timer2.Start()
+            End Using
+
+
+
         End If
+        '###### Actions if file not exists ends here
 
 
 
 
         '***** FUNCTION UPDATE ENDS HERE *****
-        If Not System.IO.File.Exists(filepath) Then
+
+        If SaveFlag = True Then
             Dim tray As New Boss
             tray.MdiParent = Kobt
             tray.Label2.Text = Me.TextBox1.Text
@@ -95,16 +123,13 @@ Public Class Timer
             Me.Close()
             tray.Show()
         End If
-
-
-        Label2.Text = Format(temp2, "HH:mm:ss")
     End Sub
 
     Private Sub Button2_Click(sender As System.Object, e As System.EventArgs) Handles Button2.Click
         Label2.Text = "00:00:00"
     End Sub
 
-    Private Sub Button3_Click_1(sender As Object, e As EventArgs) Handles Button3.Click
+    Private Sub Button3_Click_1(sender As Object, e As EventArgs)
         Dim currentuser As String = Environment.UserName
         Dim strFolder = "C:\users\" + currentuser + "\documents\kobt\"
         If String.IsNullOrEmpty(TextBox1.Text) Then
@@ -118,10 +143,8 @@ Public Class Timer
 
                 End Using
                 IO.File.WriteAllText(strFolder + TextBox1.Text, Label2.Text)
-                'MessageBox.Show("Saved as " + TextBox1.Text, "press ok", MessageBoxButtons.OK)
             Else
                 IO.File.WriteAllText(strFolder + TextBox1.Text, Label2.Text)
-                'MessageBox.Show("Saved as " + TextBox1.Text, "press ok", MessageBoxButtons.OK)
             End If
 
         End If
